added tests in services

+++ src/main/java/com/monika/worek/orchestra/service/TwoFactorAuthService.java
@@ -21,10 +21,6 @@ public class TwoFactorAuthService {
         this.verificationCodeRepository = verificationCodeRepository;
     }
 
-    public String generateVerificationCode() {
-        return String.format("%06d", new Random().nextInt(999999));
-    }
-
     public void createCode(String email, String code, int validityInMinutes) {
         verificationCodeRepository.deleteByEmail(email);
         VerificationCode verificationCode = new VerificationCode();
@@ -33,6 +29,10 @@ public class TwoFactorAuthService {
         verificationCode.setExpiryDate(LocalDateTime.now().plusMinutes(validityInMinutes));
 
         verificationCodeRepository.save(verificationCode);
+    }
+
+    private String generateVerificationCode() {
+        return String.format("%06d", new Random().nextInt(999999));
     }
 
     @Transactional
@@ -45,7 +45,9 @@ public class TwoFactorAuthService {
 
     public boolean isCodeValid(String userEmail, String userEnteredCode) {
         VerificationCode verificationCode = verificationCodeRepository.findByEmail(userEmail);
-        if (verificationCode.getCode() != null && verificationCode.getCode().equals(userEnteredCode)) {
+        if (verificationCode != null && verificationCode.getCode() != null
+                && verificationCode.getCode().equals(userEnteredCode)
+                && verificationCode.getExpiryDate().isAfter(LocalDateTime.now())) {
             verificationCodeRepository.delete(verificationCode);
             return true;
         }
